Fix Bottom mini hex border width
The mini hexagon borders are mostly double-thick. We need an extra row
of mini hexagons at the bottom to get the double border thickness on the
bottom row. The extra row won't actually be visible - we crop it to the
hexagon anyway. So the easy solution here is to just iterate fro -1
instead of 0.