Consistent mini hex wall thickness
My original approach for the mini hexagons was to put them side by side
and avoid overlaps. This works well for math calculations but leads to a
problem where some walls are thinner than others if there's no hexagon
next to it (e.g. in the center, if blank for text, and on the bottom row
if we don't add an extra row beneath it).

To fix this, I changed the approach so that the mini hexagons operate
mathematically using the concept of a wall with 0 thickness, but it
extends in both directions from that midpoint leading to a full wall
thickness for a single mini hexagon.

Along the way, I fixed a couple other minor mistakes and
inconsistencies. In one spot we were incorrectly doubling the wall
thickness (calculating a radius as if it were a diameter).

I also realized that the wall thickness needs to be applied only in the
height direction. Applying it in the width direction is incorrect -
because of the way the walls intersect at the vertices the thickness is
slightly (barely noticeable) different there.